Add files via upload
i am having a hard time to get my visual studio work because it froze. I can only take a screenshot of it , no exiting or save or anything else was available as option for me. 
I know there is solution posted online but i don't want to use that either just fort the sake of submitting it. I am wondering if I could resubmit it tomorrow morning after I get to charge my laptop and restart it to run the code again.